+++ adaptateurs/django-admin/exemple/equivalences/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class EquivalencesConfig(AppConfig):
+    name = "equivalences"
+    verbose_name = "Équivalences de diplômes"
+    default_auto_field = "django.db.models.BigAutoField"