+++ adaptateurs/django-admin/sdcd_admin/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class SdcdAdminConfig(AppConfig):
+    name = "sdcd_admin"
+    verbose_name = "Thème SDCD pour l’administration"